now using context for better component re-rendering and simplicity.

+++ src/components/SolveModal.tsx
@@ -1,10 +1,10 @@
-import React, { useState } from 'react';
+import React, { useContext } from 'react';
 import useDidMountEffect from './utils/useDidMountEffect';
 import { PuzzleSolve, WCAEvent } from '../types';
 import millisecondsToSeconds from './utils/millisecondsToSeconds';
 import { changePenaltyOfTime, deleteTime } from './utils/storageTools';
 import { BiX } from 'react-icons/bi';
-import { getTimes } from './utils/storageTools';
+import SolvesContext from './contexts/SolvesContext';
 import eventNameToFullName from './utils/eventNameToFullName';
 
 export interface SolveModalProps {
@@ -12,29 +12,19 @@ export interface SolveModalProps {
   solve: PuzzleSolve;
   isActive: boolean;
   setIsActive: any;
-  setShouldSolvesUpdate?: any; //For SolveModalForAnalytics
-  shouldSolvesUpdate?: boolean; //For SolveModalForAnalytics
 }
 
-const SolveModal = ({ eventName, solve, isActive, setIsActive, setShouldSolvesUpdate, shouldSolvesUpdate }: SolveModalProps) => { //Needs wrapper to pass in isActive, setIsActive
-  const [shouldModalUpdate, setShouldModalUpdate] = useState<boolean>(false); //Used for penalty changes
-  const [_solve, setSolve] = useState<PuzzleSolve>(solve);
-
-  useDidMountEffect(() => {
-    const updatedSolve = getTimes(eventName)?.filter((time) => time.solveId === solve.solveId);
-    console.log(updatedSolve?.[0])
-    if (updatedSolve?.[0]) {
-      setSolve(updatedSolve[0]);
-
-      if (setShouldSolvesUpdate) { //Analytics implementation
-        setShouldSolvesUpdate(!shouldSolvesUpdate)
-      }
-     }
-  }, [shouldModalUpdate]);
+const SolveModal = ({
+  eventName,
+  solve,
+  isActive,
+  setIsActive,
+}: SolveModalProps) => {
+  const { updateSolves } = useContext(SolvesContext);
 
   return (
     <div className={`modal ${isActive ? 'is-active' : ''}`}>
-      <div className="modal-background"></div>
+      <div className="modal-background" onClick={() => setIsActive(false)}></div>
       <div className="modal-content">
         <div className="box">
           <div className="is-flex is-flex-direction-column is-flex-wrap-nowrap is-justify-content-center has-background-white">
@@ -42,9 +32,9 @@ const SolveModal = ({ eventName, solve, isActive, setIsActive, setShouldSolvesUp
               <BiX color="black" size={40} onClick={() => setIsActive(false)} />
             </a>
             <h2 className="title is-2">{eventNameToFullName(eventName)}</h2>
-            <h4 className="title is-4">Time: {millisecondsToSeconds(_solve.time).toFixed(2)}</h4>
-            <h4 className="title is-4">Scramble: {_solve.scramble}</h4>
-            <h4 className="title is-4">Penalty: {_solve.penalty ? _solve.penalty.type : 'None'}</h4>
+            <h4 className="title is-4">Time: {millisecondsToSeconds(solve.time).toFixed(2) + (solve.penalty ? ` (${solve.penalty.type})`)}</h4>
+            <h4 className="title is-4">Scramble: {solve.scramble}</h4>
+            <h4 className="title is-4">Penalty: {solve.penalty ? solve.penalty.type : 'None'}</h4>
             <h4 className="title is-4">
               Timestamp:{' '}
               {new Date(solve.date).toLocaleTimeString() +
@@ -56,7 +46,7 @@ const SolveModal = ({ eventName, solve, isActive, setIsActive, setShouldSolvesUp
                 <a
                   onClick={() => {
                     changePenaltyOfTime(eventName, solve.solveId, { type: '+2', amount: 2 });
-                    setShouldModalUpdate(!shouldModalUpdate);
+                    updateSolves();
                   }}
                 >
                   +2
@@ -66,7 +56,7 @@ const SolveModal = ({ eventName, solve, isActive, setIsActive, setShouldSolvesUp
                 <a
                   onClick={() => {
                     changePenaltyOfTime(eventName, solve.solveId, { type: 'DNF' });
-                    setShouldModalUpdate(!shouldModalUpdate);
+                    updateSolves();
                   }}
                 >
                   DNF
@@ -75,11 +65,10 @@ const SolveModal = ({ eventName, solve, isActive, setIsActive, setShouldSolvesUp
               <div className="m-3 is-size-3 is-link-dark">
                 <a
                   onClick={() => {
-                    deleteTime(eventName, solve.solveId);
-                    setIsActive(false);
-
-                    if (setShouldSolvesUpdate) {
-                      setShouldSolvesUpdate(!shouldSolvesUpdate);
+                    if (window.confirm('Are you sure you want to delete your previous time?')) {
+                      deleteTime(eventName, solve.solveId);
+                      updateSolves();
+                      setIsActive(false);
                     }
                   }}
                 >
